+++ posthog-android/CHANGELOG.md
@@ -1,9 +1,10 @@
 ## Next
 
+- feat: Expose `getFeatureFlagResult` to public API ([#405](https://github.com/PostHog/posthog-android/pull/405))
+
 # 3.30.0 - 2026-01-30
 
 - update android types for new survey customization options ([#399](https://github.com/PostHog/posthog-android/pull/399))
-
 - fix: Session Replay masks only the text area for EditText, Button, and CompoundButton (CheckBox, RadioButton, Switch) instead of the entire component bounds ([#400](https://github.com/PostHog/posthog-android/pull/400))
 
 ## 3.29.1 - 2026-01-21