+++ posthog/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Next
 
+- feat: Expose `getFeatureFlagResult` to public API ([#405](https://github.com/PostHog/posthog-android/pull/405))
 - feat: Add `getFeatureFlagResult` to `PostHogFeatureFlagsInterface`, drop `getFeatureFlag`, `getFeatureFlagPayload`. ([#398](https://github.com/PostHog/posthog-android/pull/398))
 
 ## 6.3.1 - 2026-01-29